Refactor audio handling and update import paths
- Created a new `audio.py` module within the `utils/audio` directory to encapsulate audio-related functions, including thread management for playing warning sounds.
- Updated import paths across various modules to reflect the new structure, ensuring consistent access to audio functions.
- Enhanced the Python import helper with updated paths for improved code completion related to audio functionalities.
